Add TransportInfo header to common folder.
